Turn off caching for 404's

+++ app/shout_mouth.rb
@@ -155,20 +155,26 @@ class ShoutMouth < Sinatra::Base
   # Catch All
   get '/*' do
     legacy_route = LegacyRoute.first(:slug => params[:splat])
-    redirect legacy_route.post.permalink, 301 unless legacy_route.nil?
-    
-    #Dont bother caching the 404's because the webserver will not render the correct status code....
-    #Breaking the cache from the gem does not work - Manual deletion
-    cached_file = File.join(File.dirname(__FILE__), "..", "public", "cache", params[:splat])
-    cached_file += ".html" if File.extname(cached_file) == ''
-    FileUtils.rm_rf(cached_file)
-    
+    redirect legacy_route.post.permalink, 301 unless legacy_route.nil?    
     status 404
   end
-  
+
   #Default Handler In Sinatra for 404 errors....
   not_found do    
-    haml :not_found, :cache => false
+     haml :not_found
+  end
+  
+  after do
+      if response.status == 404
+        #Dont bother caching the 404's because the webserver will not render the correct status code....
+        #Breaking the cache from the gem does not work - Manual deletion
+         file = request.env["PATH_INFO"].to_s
+         file.slice!(0)
+         
+         cached_file = File.join(File.dirname(__FILE__), "..", "public", "cache", file)
+         cached_file += ".html" if File.extname(cached_file) == ''
+         FileUtils.rm_rf(cached_file)
+      end
   end
